Move serializer into dedicated package
