+++ src/main/kotlin/tech/relaycorp/relaynet/ramf/RAMFSerializer.kt
@@ -1,7 +1,7 @@
 /**
  * This class file was automatically generated by jASN1 v1.11.3-SNAPSHOT (http://www.beanit.com)
  */
-package tech.relaycorp.relaynet
+package tech.relaycorp.relaynet.ramf
 
 import com.beanit.jasn1.ber.BerLength
 import com.beanit.jasn1.ber.BerTag
